합성 애노테이션의 적용
