+++ helloboot/src/main/java/noelspring/helloboot/MySpringbootAnnotation.java
@@ -1,0 +1,16 @@
+package noelspring.helloboot;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target(ElementType.TYPE)
+@Configuration
+@ComponentScan
+public @interface MySpringbootAnnotation {
+}
